change: Use TypeScript for LoginCommand
